Make sure Digest loading is thread-safe
I'm defining the hashers in constants to avoid any performance overhead
by the mutex since we hash a lot of URLs

Digest() is only thread safe in Ruby 2.2.0 and newer. Patch:
https://github.com/ruby/ruby/commit/c02fa39463a0c6bf698b01bc610135604aca2ff4

I tried creating a test for this, but it's hard to provoke the error
with the old code to verify the test.

Closes #20