Ensure that #valid? works for protocol-less urls
There were lots of options to go with there:

* nil check in #scheme or #normalized_scheme
* .to_s in #scheme or #normalized_scheme
* use Addressable’s #normalized_scheme

I don’t think nil checks are that pretty, so that one was quickly
discarded. I also don’t think we should #to_s “randomly” for some
properties it’s either all or nothing in my opinion. Using
Addressable’s own #normalized_scheme was the top candidate, but it did
additional things to the url (not just downcasing it) and I think it’s
a good thing that we try to control what we can when it comes to
normalization.

So ultimately I went with this, I changed so that we check against
#scheme instead of #normalized_scheme as that makes a whole lot of
sense to me at least.

Fix #55.