Document the current behaviour with libidn
Currently, this URL isn't valid, but if we drop libidn (which we talked
about), it would be valid, due to the way Addressable 2.4.0 works.
See comment at: https://github.com/twingly/twingly-url/issues/66#issuecomment-246349029

Adding this URL to the tests so we will notice the change in behaviour,
if/when we drop libidn.